chore: refactor help components
