+++ src/pages/paywall/index.tsx
@@ -7,16 +7,44 @@ import { FunctionComponent } from 'react';
 import styles from './paywall.module.scss';
 import { useRouter } from 'next/router';
 import classNames from 'classnames';
-import { BylineDot } from './article/[id]';
 
-export async function getServerSideProps() {
-  return {
-    props: {
-      articles: ARTICLES_SHORTENED,
-    },
-  };
+/**
+ * Byline helper component
+ */
+
+const BylineDot = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" width="3" height="4" viewBox="0 0 3 4" fill="none">
+    <circle cx="1.5" cy="1.66016" r="1.5" fill="#434344" />
+  </svg>
+);
+
+function calculateReadingTime(text: string[], wordsPerMinute = 200) {
+  const words = text
+    .join('')
+    .split(/\s+/)
+    .filter((word) => word.length > 0);
+  const readingTimeMins = Math.round(words.length / wordsPerMinute);
+  return `${Math.max(readingTimeMins, 1)} min read`;
 }
 
+export const Byline = ({ article, includeReadingTime }: { article: ArticleData; includeReadingTime?: boolean }) => (
+  <div className={styles.byline}>
+    <Image src={article.author.avatar} className={styles.authorImage} alt={`Picture of ${article.author.name}`} />
+    <div>{article.author.name}</div>
+    <BylineDot />
+    <div>{article.date}</div>
+    {includeReadingTime && (
+      <>
+        <BylineDot />
+        <div>{calculateReadingTime(article.content)}</div>
+      </>
+    )}
+  </div>
+);
+
+/**
+ * Article Card and Grid
+ */
 type ArticleCardProps = {
   article: ArticleData;
   embed?: boolean;
@@ -32,12 +60,7 @@ const ArticleCard: FunctionComponent<ArticleCardProps> = ({ article, embed, isHe
     >
       <Image src={article.image} alt="" className={styles.articleCardImage} sizes="100vw" />
       <div className={styles.articleCardContent}>
-        <div className={styles.byline}>
-          <Image src={article.author.avatar} className={styles.authorImage} alt={`Picture of ${article.author.name}`} />
-          <div>{article.author.name}</div>
-          <BylineDot />
-          <div>{article.date}</div>
-        </div>
+        <Byline article={article} />
         <a href={link} key={article.id} className={styles.articleCardTitle}>
           {article.title}
         </a>
@@ -52,6 +75,27 @@ const ArticleCard: FunctionComponent<ArticleCardProps> = ({ article, embed, isHe
   );
 };
 
+const ArticleGrid: FunctionComponent<{ articles: ArticleData[]; embed?: boolean }> = ({ articles, embed }) => {
+  return (
+    <div className={styles.articles}>
+      {articles.map((article) => (
+        <ArticleCard key={article.id} article={article} embed={embed} />
+      ))}
+    </div>
+  );
+};
+
+/**
+ * Main Paywall use case page with article listing
+ */
+export async function getServerSideProps() {
+  return {
+    props: {
+      articles: ARTICLES_SHORTENED,
+    },
+  };
+}
+
 type PaywallProps = CustomPageProps & {
   articles: ArticleData[];
 };
@@ -62,13 +106,7 @@ export default function Paywall({ articles, embed }: PaywallProps) {
   return (
     <UseCaseWrapper useCase={USE_CASES.paywall} embed={embed} contentSx={{ maxWidth: 'none' }}>
       {heroArticle && <ArticleCard article={heroArticle} embed={embed} isHeroArticle />}
-      {gridArticles && (
-        <div className={styles.articles}>
-          {gridArticles.map((article) => (
-            <ArticleCard key={article.id} article={article} embed={embed} />
-          ))}
-        </div>
-      )}
+      {gridArticles && <ArticleGrid articles={gridArticles} embed={embed} />}
     </UseCaseWrapper>
   );
 }
